+++ src/app/pages/spending/service/spendings-sync.service.ts
@@ -1,31 +1,37 @@
 import { Injectable } from '@angular/core';
-import { Observable, tap, catchError, EMPTY, map, of } from 'rxjs';
+import { Observable, tap, catchError, EMPTY, map, of, firstValueFrom } from 'rxjs';
 import { Category } from '../../../domain/category.domain';
 import { Spending } from '../model/Spending';
-import { editSpending, addMultipleSpendings, addSpending } from '../store/spendings.actions';
+import { editSpending, addMultipleSpendings, addSpending, deleteSpending } from '../store/spendings.actions';
 import { ISpendingsState } from '../store/spendings.reducer';
 import { HttpClient } from '@angular/common/http';
 import { Store } from '@ngrx/store';
+import { OfflineStorageService } from './offline-storage.service';
 
 @Injectable({
   providedIn: 'root'
 })
 export class SpendingsSyncService {
-  private readonly url: string = 'http://localhost:8000/api/v1/profile/';
+  private readonly url: string = 'http://pegazzo.online:8000/api/v1/profile/';
 
   constructor(
     private http: HttpClient,
-    private store: Store<ISpendingsState>,
+    private store$: Store<ISpendingsState>,
+    private offlineStorage: OfflineStorageService,
   ) { }
 
   public sendSpendingToServer(portfolioID: number, spending: Spending, categories: Category[]): Observable<Spending> {
+    if(!navigator.onLine) {
+      return EMPTY;
+    }
+
     const transformedToApi = Spending.mapToSpendingApi(spending);
 
     const savedSpendingUrl = this.url + portfolioID + '/add-spending';
     return this.http.post(savedSpendingUrl, transformedToApi).pipe(
       tap((response: any) => {
         const transformedFromApi = Spending.mapFromSpendingApi(response, categories);
-        this.store.dispatch(editSpending({ spending: transformedFromApi }));
+        this.store$.dispatch(editSpending({ spending: transformedFromApi }));
       }), 
       catchError(error => {
         console.error('Error occurred while saving spending:', error);
@@ -35,6 +41,16 @@ export class SpendingsSyncService {
   }
 
   public syncSpendingListWithServer(spendingState: ISpendingsState, portfolioID: number, categories: Category[]): Observable<ISpendingsState> {
+    if(!navigator.onLine) {
+      return EMPTY;
+    }
+
+    // const failedSpendingsRequests: string[] = this.offlineStorage.getFailedSpendings();
+    // if(failedSpendingsRequests.length > 0) {
+    //   failedSpendingsRequests.forEach(spendingId => this.store$.dispatch(deleteSpending({id: spendingId})));
+    //   console.log('failedSpendingsRequests',failedSpendingsRequests);
+    // }
+    
     const loadSpendingsUrl = this.url + 'spendings-list/' + portfolioID;
     return this.http.get<Spending[]>(loadSpendingsUrl).pipe(
       map(serverSpendings => {
@@ -42,7 +58,7 @@ export class SpendingsSyncService {
         const newSpendingsFromServer = this.filterNewSpendingsFromServer(serverSpendings, clientSpendings, categories);
         
         if (newSpendingsFromServer.length > 0) {
-          this.store.dispatch(addMultipleSpendings({ spendings: newSpendingsFromServer }));
+          this.store$.dispatch(addMultipleSpendings({ spendings: newSpendingsFromServer }));
         }
 
         this.sendUnsavedSpendingsToServer(clientSpendings);
@@ -57,6 +73,11 @@ export class SpendingsSyncService {
   }
 
   public deleteSpending(spendingId: string): Observable<void> {
+    if(!navigator.onLine) {
+      this.offlineStorage.offlineDeleteSpending(spendingId);
+      return EMPTY;
+    }
+
     const deleteUrl = this.url + 'delete-spending/' + spendingId;
 
     return this.http.delete(deleteUrl).pipe(
@@ -75,7 +96,7 @@ export class SpendingsSyncService {
     clientSpendings
       .filter(spending => !spending.isSaved)
       .forEach(unsavedSpending => {
-        this.store.dispatch(addSpending({ spending: unsavedSpending }));
+        this.store$.dispatch(addSpending({ spending: unsavedSpending }));
       });
   }
 }
